Unconditionally initialize the POSIX semaphore supporting a dispatch semaphore, as the type sem_t may not be a scalar on all platforms.
Even if it were, the current code was racey as it failed to do an
atomic test-and-set to make sure that two threads hadn't triggered
on-demand initialization of the dispatch semaphore at the same time.

In practice, this should remain fairly efficient as non-pshared POSIX
semaphores will generally be implemented entirely with libpthread,
and likely not require kernel involvement during allocation (as is
required for Mach semaphores, hence lazy allocation optimizations).

Spotted by:	Kevin Van Vechten <kvv@apple.com>



git-svn-id: http://svn.macosforge.org/repository/libdispatch/trunk@84 5710d607-3af0-45f8-8f96-4508d4f60227