> Mark Heily: > > I have attached a collection of patches that allows libdispatch to build on > Linux and Solaris. Many thanks to Joakim Johansson for providing the Solaris > patches and testing everything. > > The commands I used to build on Linux are: > >   CC="clang" LD="llvm-ld-2.8" CFLAGS="-I/usr/local" \ >   LDFLAGS="-L/usr/local" LIBS="-lbsd" ./configure && make > > The commands Joakim used to build on Solaris are: > >   ./configure CFLAGS="-m64 -I/usr/local/include" \ >   LDFLAGS="-L/usr/local/lib” gmake
My Solaris and Linux environments are limping at the moment (dev tools all
horked up in both) so I couldn't do the build tests on anything but Darwin,
but the patches pass the eyeball check in any case.  Also made a small
change to the xcode project to increase buildability (no legacy APIs).



git-svn-id: http://svn.macosforge.org/repository/libdispatch/trunk@194 5710d607-3af0-45f8-8f96-4508d4f60227